+++ changelog.md
@@ -1,2 +1,3 @@
 - Add activation ratelimiter
 - Add reversal key for cycling macros
+- Add placeholder for targeted block